feat: #86 Move discord enforcement code somewhere else to simplify the scanner service.
